requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.2
idna==2.9
multidict==4.7.5
pip==19.2.3
pytz==2020.1
setuptools==41.2.0
websockets==8.1
yarl==1.4.2